code clean up

+++ src/Moonglade.Web/Controllers/ImageController.cs
@@ -114,9 +114,9 @@ public class ImageController : ControllerBase
                 watermarkedStream = watermarker.AddWatermark(
                     _blogConfig.ImageSettings.WatermarkText,
                     Color.FromRgba(
-                        (byte)128,
-                        (byte)128,
-                        (byte)128,
+                        128,
+                        128,
+                        128,
                         (byte)_blogConfig.ImageSettings.WatermarkColorA),
                     WatermarkPosition.BottomRight,
                     15,
